Website Update
Created an article element to contain information about the game and created an aside element to contain a list of customer reviews. Created a variable to store HTML for a table displaying the product, price, platform, esrb rating, condition, release, and summary of an item and writes that content to the article element.